+++ pkg/vm/engine/aoe/storage/layout/index/numbsi.go
@@ -77,24 +77,45 @@ func (i *NumericBsiIndex) GetCol() int16 {
 }
 
 func (i *NumericBsiIndex) Eval(ctx *FilterCtx) error {
+	if ctx.BMRes.IsEmpty() {
+		return nil
+	}
 	var err error
 	if v, ok := ctx.Val.(types.Date); ok {
 		ctx.Val = int32(v)
+		defer func() {
+			ctx.Val = v
+		}()
 	}
 	if v, ok := ctx.ValMin.(types.Date); ok {
 		ctx.ValMin = int32(v)
+		defer func() {
+			ctx.ValMin = v
+		}()
 	}
 	if v, ok := ctx.ValMax.(types.Date); ok {
 		ctx.ValMax = int32(v)
+		defer func() {
+			ctx.ValMax = v
+		}()
 	}
 	if v, ok := ctx.Val.(types.Datetime); ok {
 		ctx.Val = int64(v)
+		defer func() {
+			ctx.Val = v
+		}()
 	}
 	if v, ok := ctx.ValMin.(types.Datetime); ok {
 		ctx.ValMin = int64(v)
+		defer func() {
+			ctx.ValMin = v
+		}()
 	}
 	if v, ok := ctx.ValMax.(types.Datetime); ok {
 		ctx.ValMax = int64(v)
+		defer func() {
+			ctx.ValMax = v
+		}()
 	}
 	switch ctx.Op {
 	case OpEq:
@@ -343,7 +364,7 @@ func BuildNumericBsiIndex(data []*vector.Vector, t types.Type, colIdx int16) (bs
 		for _, part := range data {
 			column := part.Col.([]types.Datetime)
 			for _, val := range column {
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
+				if err := bsiIdx.Set(uint64(row), int64(val)); err != nil {
 					return nil, err
 				}
 				row++
@@ -356,7 +377,7 @@ func BuildNumericBsiIndex(data []*vector.Vector, t types.Type, colIdx int16) (bs
 		for _, part := range data {
 			column := part.Col.([]types.Date)
 			for _, val := range column {
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
+				if err := bsiIdx.Set(uint64(row), int32(val)); err != nil {
 					return nil, err
 				}
 				row++
