+++ pkg/vm/engine/aoe/storage/options.go
@@ -39,6 +39,7 @@ const (
 
 	DefaultBlockWriters     = uint16(8)
 	DefaultSegmentWriters   = uint16(4)
+	DefaultIndexWriters     = uint16(4)
 	DefaultStatelessWorkers = uint16(1)
 )
 
@@ -108,6 +109,7 @@ func (o *Options) FillDefaults(dirname string) *Options {
 		o.SchedulerCfg = &SchedulerCfg{
 			BlockWriters:     DefaultBlockWriters,
 			SegmentWriters:   DefaultSegmentWriters,
+			IndexWriters: DefaultIndexWriters,
 			StatelessWorkers: DefaultStatelessWorkers,
 		}
 	} else {
@@ -116,6 +118,9 @@ func (o *Options) FillDefaults(dirname string) *Options {
 		}
 		if o.SchedulerCfg.SegmentWriters == 0 {
 			o.SchedulerCfg.SegmentWriters = DefaultSegmentWriters
+		}
+		if o.SchedulerCfg.IndexWriters == 0 {
+			o.SchedulerCfg.IndexWriters = DefaultIndexWriters
 		}
 		if o.SchedulerCfg.StatelessWorkers == 0 {
 			o.SchedulerCfg.StatelessWorkers = DefaultStatelessWorkers
